Added optional second argument for SSR defaults
If matchMedia is not present, then it MAY be the case that we're
doing a server-side render. In that case the previous behavior of
just returning the empty set may not always be useful, so we will
give users a way to specify a default return.

Fixed up the tests to exercise those cases.